@@ -13,15 +13,15 @@
     }
   ],
   "require": {
-    "php": ">=7.1",
+    "php": ">=7.2",
     "ext-pdo": "*",
     "ext-json": "*",
-    "twig/twig":"~2.0",
-    "phpunit/phpunit": "^6.0",
-    "predis/predis":"1.1.*@dev",
-    "codacy/coverage": "dev-master",
-    "auth0/auth0-php": "^5.0",
-    "ext-json": "*"
+    "ext-openssl": "*",
+    "twig/twig":"~3.0",
+    "phpunit/phpunit": "^8.0",
+    "predis/predis":"^1.1",
+    "auth0/auth0-php": "^7.4",
+    "ext-ldap": "*"
   },
   "suggest": {
     "ext-ctype": "Provide ctype function support for installs without the ctype extension."
@@ -30,8 +30,5 @@
     "psr-0": {
       "Staple": "library/"
     }
-  },
-  "require-dev": {
-    "codacy/coverage": "dev-master"
   }
 }
